cleanup (cli): move tests/ directory to agenta/agenta-cli/agenta
