+++ agenta-cli/agenta/sdk/__init__.py
@@ -20,7 +20,9 @@ from .decorators.tracing import instrument
 from .decorators.llm_entrypoint import entrypoint, app, route
 from .agenta_init import Config, AgentaSingleton, init
 from .utils.helper.openai_cost import calculate_token_usage
-from .config_manager import ConfigManager
+from .managers.config_manager import ConfigManager
+from .managers.variant_manager import VariantManager
+from .managers.deployment_manager import DeploymentManager
 
 config = PreInitObject("agenta.config", Config)
 DEFAULT_AGENTA_SINGLETON_INSTANCE = AgentaSingleton()
