+++ agenta-cli/agenta/sdk/managers/__init__.py
@@ -1,6 +1,6 @@
-from .config_manager import ConfigManager
-from .variant_manager import VariantManager
-from .deployment_manager import DeploymentManager
+from agenta.sdk.managers.config_manager import ConfigManager
+from agenta.sdk.managers.variant_manager import VariantManager
+from agenta.sdk.managers.deployment_manager import DeploymentManager
 
 
 __all__ = ["ConfigManager", "VariantManager", "DeploymentManager"]
