+++ agenta-cli/agenta/tests/management_sdk/manual_tests/apps_with_new_sdk/new_chat_prompt.py
