+++ agenta-cli/agenta/tests/management_sdk/manual_tests/apps_with_new_sdk/run_local.py
