+++ agenta-cli/agenta/tests/management_sdk/manual_tests/apps_with_new_sdk/v3.py
