+++ agenta-cli/agenta/tests/management_sdk/manual_tests/apps_with_old_sdk/chat_prompt.py
