+++ agenta-cli/agenta/tests/management_sdk/v2.py
